WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/017.html

<!doctype html>
<title>WebSockets: this, e.target, e.currentTarget, e.eventPhase</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo_raw');
  ws.addEventListener('open', function(e) {
    
    // first a text frame, then a frame with reserved opcode 3
    // which should fail the connection
    ws.send('\\x81\\x04test\\x83\\x03LOL');
    assert_equals(this, ws);
    assert_equals(e.target, ws);
    assert_equals(e.currentTarget, ws);
    assert_equals(e.eventPhase, 2);
  }, false);
  ws.addEventListener('message', function(e) {
    
    assert_equals(this, ws);
    assert_equals(e.target, ws);
    assert_equals(e.currentTarget, ws);
    assert_equals(e.eventPhase, 2);
  }, false);
  ws.addEventListener('error', function(e) {
    
    assert_equals(this, ws);
    assert_equals(e.target, ws);
    assert_equals(e.currentTarget, ws);
    assert_equals(e.eventPhase, 2);
  }, false);
  ws.addEventListener('close', function(e) {
    
    assert_equals(this, ws);
    assert_equals(e.target, ws);
    assert_equals(e.currentTarget, ws);
    assert_equals(e.eventPhase, 2);
    end();
  }, false);  
});
</script>
